Добавлена политика AnswerAttachmentPolicy.

+++ app/controllers/api/v1/answer_attachments_controller.rb
@@ -3,6 +3,7 @@ module Api
     class AnswerAttachmentsController < BaseController
       def show
         attachment = Answer.find(params[:answer_id]).attachments.find(params[:id])
+        authorize attachment
 
         send_file attachment.document.file.file
       rescue ActionController::MissingFile
@@ -11,6 +12,7 @@ module Api
 
       def create
         attachment = Answer.find(params[:answer_id]).attachments.new(attachments_params)
+        authorize attachment
 
         if attachment.save
           render json: attachment
@@ -21,6 +23,7 @@ module Api
 
       def destroy
         attachment = Answer.find(params[:answer_id]).attachments.find(params[:id])
+        authorize attachment
 
         if attachment.destroy
           render json: attachment
